T46: extract WatchModeController from ParsekFlight
Move camera-follow / watch-mode code (847 lines, 15 fields, 18 methods)
into its own WatchModeController class. This includes the #region Camera
Follow block plus supporting methods (UpdateWatchModeValid,
UpdateWatchEndHoldTimer, UpdateWatchRetargetAfterExplosion, etc.).

ParsekFlight keeps thin forwarding methods/properties for external
callers (ParsekUI, ParsekPlaybackPolicy, tests). WatchModeController
accesses ParsekFlight state via internal accessors.
